Give some more options to the user

+++ Manage Conky/ConkyPreferencesSheetController.m
@@ -11,7 +11,6 @@
 #include <unistd.h>
 #include <ServiceManagement/ServiceManagement.h>
 
-
 @implementation ConkyPreferencesSheetController
 
 @synthesize runConkyAtStartupCheckbox = _runConkyAtStartupCheckbox;
@@ -25,27 +24,21 @@ NSString * kConkyAgentPlistName = @"org.npyl.conky.plist";
 {
     NSString * conkyAgentPlistPath = [[NSString alloc] initWithFormat:@"%@%@%@%@", @"/Users/", NSUserName(), @"/Library/LaunchAgents/", kConkyAgentPlistName ];
     
-    
     [super activateSheet:@"ConkyPreferences"];
     
     /*
-     *  Conky agent present?
+     *  Is conky agent present?
      */
     int res = access( [conkyAgentPlistPath UTF8String], R_OK);
-    if (res < 0) {
+    if (res < 0)
         NSLog( @"Agent plist doesnt exist or not accessible!" );
-        
-        // NOTE: by default the checkbox is unchecked
-    } else {
+    else
         [_runConkyAtStartupCheckbox setState:1];
-    }
-    
     
     /*
      *  Conky configuration file location?
      */
     NSString * conkyConfigsPath = [[NSUserDefaults standardUserDefaults] objectForKey:@"configsLocation"];
-    
     if (!conkyConfigsPath)
     {
         NSString * kConkyConfigsDefaultPath = [NSString stringWithFormat:@"/Users/%@/.conky", NSUserName()];
@@ -58,7 +51,8 @@ NSString * kConkyAgentPlistName = @"org.npyl.conky.plist";
     }
     
     [_conkyConfigLocationTextfield setStringValue:conkyConfigsPath];
-    [_conkyConfigLocationTextfield setDelegate:self];   /* Do that to allow, getting the Enter-Key notification */
+    /* Do that to allow, getting the Enter-Key notification */
+    [_conkyConfigLocationTextfield setDelegate:self];
 }
 
 -(void)controlTextDidEndEditing:(NSNotification *)notification
@@ -73,38 +67,48 @@ NSString * kConkyAgentPlistName = @"org.npyl.conky.plist";
 
 - (IBAction)runConkyAtStartupCheckboxAction:(id)sender
 {
+    // XXX add support for themes
+    
     NSString * conkyAgentPlistPath = [[NSString alloc] initWithFormat:@"%@%@%@%@", @"/Users/", NSUserName(), @"/Library/LaunchAgents/", kConkyAgentPlistName ];
     
-    NSString * kConkyLaunchAgentLabel = @"org.npyl.conky";
-    NSString * kConkyExecutablePath = @"/Applications/ConkyX.app/Contents/Resources/conky";
-    
-    // TODO: add support for theme
-    
-    id objects[] = { kConkyLaunchAgentLabel, @[ kConkyExecutablePath ], [NSNumber numberWithBool:YES], [NSNumber numberWithBool:NO] };
-    id keys[] = { @"Label", @"ProgramArguments", @"RunAtLoad", @"KeepAlive"  };
-    NSUInteger count = sizeof(objects) / sizeof(id);
-    
-    NSDictionary * conkyAgentPlist = [[NSDictionary alloc] initWithObjects:objects forKeys:keys count:count];
-    
-    
-    switch ([sender state]) {
-        case 0:
-            NSLog( @"Request to remove the Agent!" );
-            
-            // SMJobRemove() deprecated but suggested by Apple, see https://lists.macosforge.org/pipermail/launchd-dev/2016-October/001229.html
-            SMJobRemove( kSMDomainUserLaunchd, CFSTR("org.npyl.conky"), nil, YES, nil);
-            
-            unlink( [conkyAgentPlistPath UTF8String] );
-            
-            break;
-        case 1:
-            NSLog( @"Request to add the Agent!" );
- 
-            [conkyAgentPlist writeToFile:conkyAgentPlistPath atomically:YES];
-            
-            break;
-        default:
-            ;
+    if ([sender state] == 0)
+    {
+        NSLog( @"Request to remove the Agent!" );
+        
+        // SMJobRemove() deprecated but suggested by Apple, see https://lists.macosforge.org/pipermail/launchd-dev/2016-October/001229.html
+        SMJobRemove(kSMDomainUserLaunchd, CFSTR("org.npyl.conky"), nil, YES, nil);
+        
+        unlink( [conkyAgentPlistPath UTF8String] );
+    }
+    else if ([sender state] == 1)
+    {
+        NSLog( @"Request to add the Agent!" );
+        
+        NSString * kConkyLaunchAgentLabel = @"org.npyl.conky";
+        NSString * kConkyExecutablePath = @"/Applications/ConkyX.app/Contents/Resources/conky";
+        
+        id objects[] = { kConkyLaunchAgentLabel, @[ kConkyExecutablePath ], [NSNumber numberWithBool:YES], [NSNumber numberWithBool:YES] };
+        id keys[] = { @"Label", @"ProgramArguments", @"RunAtLoad", @"KeepAlive" };
+        NSUInteger count = sizeof(objects) / sizeof(id);
+        
+        NSDictionary * conkyAgentPlist = [[NSDictionary alloc] initWithObjects:objects forKeys:keys count:count];
+        
+        NSAlert *keepAlivePrompt = [[NSAlert alloc] init];
+        [keepAlivePrompt setMessageText:@"Select your preference"];
+        [keepAlivePrompt setInformativeText:@"Always restart conky when for some reason it quits?"];
+        [keepAlivePrompt setAlertStyle:NSAlertStyleInformational];
+        [keepAlivePrompt addButtonWithTitle:@"YES"];
+        [keepAlivePrompt addButtonWithTitle:@"NO"];
+        
+        NSModalResponse response = [keepAlivePrompt runModal];
+        switch (response)
+        {
+            case NSAlertSecondButtonReturn:
+                objects[3] = [NSNumber numberWithBool:NO];
+                break;
+        }
+        
+        [conkyAgentPlist writeToFile:conkyAgentPlistPath atomically:YES];
     }
 }
